requirements.txt
torch
pandas
matplotlib
pillow
torchvision
openpyxl
seaborn